pyproject.toml
[tool.poetry]
name = "simple-sftp"
version = "0.1"
description = "Simple SFTP python client based on ssh2-python package"
authors = ["Ivan Fedorov <inbox@titaniumhocker.ru>"]
license = "MIT"
readme = "README.rst"
homepage = "https://github.com/TitaniumHocker/simple-sftp"
repository = "https://github.com/TitaniumHocker/simple-sftp"
documentation = "https://simple-sftp.rtfd.io"
keywords = [
    "sftp", "ssh"
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python",
    "Programming Language :: Python :: 3.6",
    "Programming Language :: Python :: 3.7",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Intended Audience :: Developers",
    "Operating System :: POSIX :: Linux",
    "Topic :: Internet",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development :: Libraries :: Python Modules"
]

[tool.poetry.urls]
"Bug Tracker" = "https://github.com/TitaniumHocker/simple-sftp/issues"

[tool.poetry.dependencies]
python = "^3.6"
ssh2-python = "^0.26.0"

[tool.poetry.dev-dependencies]
tox = "^3.21.4"
mypy = "^0.812"
pytest = "^6.2.2"
flake8 = "^3.8.4"
pytest-cov = "^2.11.1"
coverage = {extras = ["toml"], version = "^5.4"}
flake8-isort = "^4.0.0"
flakehell = "^0.9.0"
flake8-docstrings = "^1.5.0"
factory-boy = "^3.2.0"
pytest-factoryboy = "^2.1.0"
flake8-black = "^0.2.1"
jedi = "^0.18.0"
isort = "^5.7.0"
black = "^20.8b1"
paramiko = "^2.7.2"
cryptography = "^3.4.6"
mock-ssh-server = "^0.8.2"
Jinja2 = "^2.11.3"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"

[tool.flakehell]
format = "colored"
max_line_length = 100

[tool.flakehell.plugins]
pyflakes = ["+*"]
pycodestyle = ["+*", "-W503"]
flake8-docstrings = ["+*", "-D400", "-D107", "-D204", "-D105"]
flake8-isort = ["+*"]
flake8-black = ["+*"]

[tool.flakehell.exceptions."tests/*"]
flake8-docstrings = ["-*"]
pycodestyle = ["-F401"]

[tool.isort]
profile = "hug"
src_paths = ["simple_sftp", "tests"]

[tool.pytest.ini_options]
testpaths = [
    "tests"
]

[tool.coverage.run]
source = [
    "simple_sftp"
]

[tool.coverage.report]
fail_under = 50
show_missing = true

[tool.tox]
legacy_tox_ini = """
[tox]
envlist = py36,py37,py38,py39
isolated_build = True

[tox:.package]
basepython = python3

[testenv]
commands = pytest
deps =
    pytest
    pytest-sftpserver
    pytest-factoryboy
"""
